refactor: change expression by request
