+++ Sources/CubbyTower/Systems/AttackSystem.cpp
@@ -5,11 +5,11 @@
 // property of any third parties.
 
 #include <CubbyTower/Commons/Tags.hpp>
-#include <CubbyTower/Components/Position.hpp>
 #include <CubbyTower/Components/AttackRange.hpp>
-#include <CubbyTower/Components/TargetMask.hpp>
 #include <CubbyTower/Components/Damage.hpp>
-#include <CubbyTower/Components/HealthPoint.hpp>
+#include <CubbyTower/Components/Health.hpp>
+#include <CubbyTower/Components/Position.hpp>
+#include <CubbyTower/Components/TargetMask.hpp>
 #include <CubbyTower/Components/TypeMask.hpp>
 #include <CubbyTower/Systems/AttackSystem.hpp>
 #include <iostream>
@@ -24,9 +24,12 @@ void Attack(entt::registry& registry)
         {
             const auto towerPos = registry.get<Position>(tower);
             const auto enemyPos = registry.get<Position>(enemy);
-            const auto attackRange = registry.get<AttackRange>(tower).attackRange;
+            const auto attackRange =
+                registry.get<AttackRange>(tower).attackRange;
 
-            if ((towerPos.x - enemyPos.x) * (towerPos.x - enemyPos.x) + (towerPos.y - enemyPos.y) * (towerPos.y - enemyPos.y) > attackRange * attackRange)
+            if ((towerPos.x - enemyPos.x) * (towerPos.x - enemyPos.x) +
+                    (towerPos.y - enemyPos.y) * (towerPos.y - enemyPos.y) >
+                attackRange * attackRange)
             {
                 continue;
             }
@@ -40,7 +43,7 @@ void Attack(entt::registry& registry)
             {
                 // TODO: Consider the target priority, make a projectile
                 // Simple attack test
-                auto& health = registry.get<HealthPoint>(enemy);
+                auto& health = registry.get<Health>(enemy);
                 health.health -= damage;
             }
         }
