+++ Tests/UnitTests/Systems/AttackSystemTests.cpp
@@ -11,7 +11,7 @@
 #include <CubbyTower/Components/Damage.hpp>
 #include <CubbyTower/Components/Distance.hpp>
 #include <CubbyTower/Components/Gold.hpp>
-#include <CubbyTower/Components/HealthPoint.hpp>
+#include <CubbyTower/Components/Health.hpp>
 #include <CubbyTower/Components/Name.hpp>
 #include <CubbyTower/Components/Position.hpp>
 #include <CubbyTower/Components/TargetMask.hpp>
@@ -26,8 +26,8 @@ using namespace CubbyTower;
 
 TEST_CASE("[AttackSystem] - Attack")
 {
-    //Testing only Attack functions running right way
-    //Todo : Real attack
+    // Testing only Attack functions running right way
+    // Todo : Real attack
     entt::registry registry;
 
     auto entity = registry.create();
@@ -38,22 +38,22 @@ TEST_CASE("[AttackSystem] - Attack")
 
     auto enemy1 = registry.create();
     registry.emplace<Tag::Enemy>(enemy1);
-    registry.emplace<HealthPoint>(enemy1, 2);
-    registry.emplace<TypeMask>(enemy1, 0b010); // Type: ground
+    registry.emplace<Health>(enemy1, 2);
+    registry.emplace<TypeMask>(enemy1, 0b010);  // Type: ground
     registry.emplace<Position>(enemy1, 0.0, 0.0);
     registry.emplace<Distance>(enemy1, 250.0);
 
     auto enemy2 = registry.create();
     registry.emplace<Tag::Enemy>(enemy2);
-    registry.emplace<HealthPoint>(enemy2, 2);
-    registry.emplace<TypeMask>(enemy2, 0b011); // Type: ground & stealth
+    registry.emplace<Health>(enemy2, 2);
+    registry.emplace<TypeMask>(enemy2, 0b011);  // Type: ground & stealth
     registry.emplace<Position>(enemy2, 0.0, 0.0);
     registry.emplace<Distance>(enemy2, 350.0);
 
     auto enemy3 = registry.create();
     registry.emplace<Tag::Enemy>(enemy3);
-    registry.emplace<HealthPoint>(enemy3, 2);
-    registry.emplace<TypeMask>(enemy3, 0b010); // Type: ground
+    registry.emplace<Health>(enemy3, 2);
+    registry.emplace<TypeMask>(enemy3, 0b010);  // Type: ground
     registry.emplace<Position>(enemy3, 0.0, 0.0);
     registry.emplace<Distance>(enemy3, 100.0);
 
@@ -61,19 +61,20 @@ TEST_CASE("[AttackSystem] - Attack")
 
     Attack(registry);
 
-    for (auto [enemy, dist, hp, mask] : registry.view<Tag::Enemy, Distance, HealthPoint, TypeMask>().each())
+    for (auto [enemy, dist, hp, mask] :
+         registry.view<Tag::Enemy, Distance, Health, TypeMask>().each())
     {
         if (mask.typeMask == 0b011)
         {
-            CHECK_EQ(hp.hp, 2);
+            CHECK_EQ(hp.health, 2);
         }
         else if (dist.distance == 250)
         {
-            CHECK_EQ(hp.hp, 1);
+            CHECK_EQ(hp.health, 1);
         }
         else if (dist.distance == 100)
         {
-            CHECK_EQ(hp.hp, 2);
+            CHECK_EQ(hp.health, 2);
         }
     }
 }
